lyft_l5_tl2la/requirements.txt
bokeh==2.3.3
wheel==0.38.4
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jupyter==1.0.0
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
jupytext==1.13.8
l5kit==1.5.0
loguru==0.6.0
numpy==1.19.5
pandas==1.4.4
plotly==5.7.0
prettytable==2.5.0
prompt-toolkit==3.0.30
protobuf==3.19.4
PTable==0.9.2
scipy==1.5.4
seaborn==0.11.2
Shapely==1.8.1.post1
toml==0.10.2
tqdm==4.64.0
zarr==2.8.3
pydotplus
pyinterval==1.2.0